Layout de registro 2.0

+++ Projeto Integrador 2/app/src/main/java/com/example/projetonothungry/model/Produtos.kt
@@ -9,7 +9,7 @@ data class Produtos(
     var descricao : String,
     var imagem : String,
     var quantidade : Int,
-    var valor : Double,
+    var valor : Int,
     var categoria : Categoria
